requirements.txt
pytchat==0.5.5
obs-websocket-py==1.0